// frontend/src/app/models/course-management/courseMatchingData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CourseMatchingData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** FK (course_main) */
    courseId?: number;
    /** PK */
    courseMatchingId?: number;
    /** ชื่อหลักสูตร (ภาษาอังกฤษ) */
    curriculumNameEn?: string;
    /** ชื่อหลักสูตร (ภาษาไทย) */
    curriculumNameTh?: string;
    /** FK swu_curriculum */
    curriculumSwuId?: string;
    /** อีเมล */
    externalEmail?: string;
    /** ชื่ออาจารย์/ผู้รับผิดชอบ  (ภาษาอังกฤษ) */
    externalNameEn?: string;
    /** ชื่ออาจารย์/ผู้รับผิดชอบ (ภาษาไทย) */
    externalNameTh?: string;
    /** FileName */
    filename?: string;
    first?: number;
    /** บุคลากรภายนอก */
    instructorType?: boolean;
    mode?: string;
    /** Module */
    module?: string;
    /** Prefix */
    prefix?: string;
    rowNum?: number;
    size?: number;
    /** รหัสวิชา (ภาษาอังกฤษ) */
    subjectCodeEn?: string;
    /** รหัสวิชา (ภาษาไทย) */
    subjectCodeTh?: string;
    /** จํานวนหน่วยกิต */
    subjectCredit?: number;
    /** ชื่อรายวิชา (ภาษาอังกฤษ) */
    subjectNameEn?: string;
    /** ชื่อรายวิชา (ภาษาไทย) */
    subjectNameTh?: string;
    /** FK (swu_subject) */
    subjectSwuId?: number;

    chairmanId?: number ;
	ownerDepCode?: string ;
	ownerDepNameTh?: string ;
	ownerDepNameEn?: string ;
	ownerFacultyCode?: string ;
	ownerFacultyNameTh?: string ;
	ownerFacultyNameEn?: string ;
	subjectSet?: string;

}