// frontend/src/app/modules/component/course-preview/course-main/general/view-course-comparison/view-course-comparison.component.ts
import { Component, Input, OnInit } from '@angular/core';
import { TranslateService } from '@ngx-translate/core';
import { NgxUiLoaderService } from 'ngx-ui-loader';
import { MessageService } from 'primeng/api';
import { TablePageEvent } from 'primeng/table';
import { CourseMainData, CourseMatchingData, SwuCurriculumData } from 'src/app/models/course-management';
import { CourseManagementService } from 'src/app/services/course-management.service';
import { DropdownService } from 'src/app/services/dropdown.service';

@Component({
    selector: 'app-view-course-comparison',
    templateUrl: './view-course-comparison.component.html',
    styleUrls: ['./view-course-comparison.component.scss']
})
export class ViewCourseComparisonComponent implements OnInit {
    @Input() lang: string;
    @Input() courseMain: CourseMainData;

    initForm: boolean = false;
    visible: boolean = false;

    items: CourseMatchingData[] = [];
    totalRecords: number = 0;

    detailSubject: SwuCurriculumData[] = [];


    constructor(
        public translate: TranslateService,
        private dropdownService: DropdownService,
        private loaderService: NgxUiLoaderService,
        private courseManagementService: CourseManagementService,
        private messageService: MessageService
    ) {}

    ngOnInit(): void {
        this.fetchData();
    }

    fetchData(event?: TablePageEvent) {
        this.loaderService.start();

        const criteria: CourseMatchingData = {
            courseId: this.courseMain.courseId,
            activeFlag: true,
            first: 0,
            size: 5,
            mode: 'search'
        };

        if (event) {
            criteria.size = event.rows;
            criteria.first = event.first;
        }

        this.courseManagementService.findCourseMatching(criteria).subscribe((result) => {
            this.loaderService.stop();
            if (result.status === 200) {
                this.items = result.entries;
                this.totalRecords = result.totalRecords;
            } else {
                this.messageService.add({
                    severity: 'error',
                    summary: this.translate.instant('common.alert.fail'),
                    detail: this.translate.instant(result.message),
                    life: 2000
                });
            }
        });
    }

    showDetail(subId?: number, curId?: string) {
        console.log('subId :>> ', subId);
        console.log('curId :>> ', curId);
        this.visible = true;

        const criteria: SwuCurriculumData = {
            subjectSwuId: subId,
            curriculumSwuId: curId
        }

        this.courseManagementService
        .findSwuCurriculum(criteria)
        .subscribe(({ status, message, entries, totalRecords }) => {
            this.loaderService.stop();
            if (status === 200) {
                this.detailSubject = entries;
                console.log('this.detailSubject :>> ', this.detailSubject);
                this.totalRecords = totalRecords;
            } else {
                this.messageService.add({
                    severity: 'error',
                    summary: this.translate.instant('common.alert.fail'),
                    detail: this.translate.instant(message),
                    life: 2000
                });
            }
        });

    }

}
